esl_sq_ReverseComplement(): fixed type in an ESL_SWAP call of an ESL_SQ's <sq->start> and <sq->end>. Change from incorrect 'int' type to 'int64_t'.
Sean please verify this is a valid correction -- there's no reason
ESL_SWAP can't handle type 'int64_t', right?


git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/trunk@937 db3e70e4-12e1-0310-b23f-81a6e3d09ba7

+++ esl_sq.c
@@ -1530,7 +1530,7 @@ esl_sq_ReverseComplement(ESL_SQ *sq)
     }
 #endif /*eslAUGMENT_ALPHABET*/
 
-  ESL_SWAP(sq->start, sq->end, int);
+  ESL_SWAP(sq->start, sq->end, int64_t);
   /* revcomp invalidates any secondary structure annotation */
   if (sq->ss != NULL) { free(sq->ss); sq->ss = NULL; }
   /* revcomp invalidates any extra residue markup */
